pyproject.toml
[tool.poetry]
name = "csat-backend"
version = "0.1.0"
description = "csat project rezonica"
authors = ["Raj Livingston"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.129.0"
sqlalchemy = "^2.0.46"
pydantic = "^2.12.5"
gunicorn = "^25.0.3"
uvicorn = "^0.40.0"
pymysql = "^1.1.2"
python-jose = "^3.5.0"
passlib = "^1.7.4"
python-multipart = "^0.0.22"
boto3 = "^1.42.48"
python-dotenv = "^1.2.1"
cryptography = "^46.0.5"
email-validator = "^2.3.0"
requests = "^2.32.5"
httpx = "^0.28.1"
pandas = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
